Add support for parsing embedded href="~/Urls" via templateBase.ResolveUrl().
Not a great solution but it attempts to either create a relative path to the template based on the root folder, or if that's not possible a `file:///` based Url.

Implemented in order to not break code, but wouldn't recommend using ~/ paths for external assets as there's no clean way to resolve the base virtual path.